requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
asyncio-mqtt==0.13.0
aiofiles==23.2.1
httpx==0.25.2